third_party/WebKit/LayoutTests/http/tests/priority-hints/fetch-api-importance.html

<script src=resources/common.js></script>
<script src=../resources/testharness.js></script>
<script src=../resources/testharnessreport.js></script>

<script>
  const resource_base = '../resources/dummy.txt';

  const tests = [
    // fetch("url", request_init) tests
    {description: 'fetch() with URL and request_init whose importance is "high" must have no effect on resource load priority', request_info: `${resource_base}?1`, request_init: {importance: 'high'}, expected_priority: kHigh},
    {description: 'fetch() with URL and request_init whose importance is "auto" must have no effect on resource load priority', request_info: `${resource_base}?2`, request_init: {importance: 'auto'}, expected_priority: kHigh},
    {description: 'fetch() with URL and request_init whose importance is missing must have no effect on resource load priority', request_info: `${resource_base}?3`, request_init: {}, expected_priority: kHigh},
    {description: 'fetch() with URL and request_init whose importance is "low" must be fetched with kLow resource load priority', request_info: `${resource_base}?4`, request_init: {importance: 'low'}, expected_priority: kLow},
    // fetch(Request, request_init) tests
    {
      description: 'fetch() with Request whose importance is "low" and request_init whose importance is "high" must have no effect on resource load priority',
      request_info: new Request(`${resource_base}?5`, {importance: 'low'}),
      request_init: {importance: 'high'},
      expected_priority: kHigh
    },
    {
      description: 'fetch() with Request whose importance is "high" and request_init whose importance is "low" must be fetched with kLow resource load priority',
      request_info: new Request(`${resource_base}?6`, {importance: 'high'}),
      request_init: {importance: 'low'},
      expected_priority: kLow
    },
    // fetch(Request) tests
    {
      description: 'fetch() with Request whose importance is "high" and no request_init must have no effect on resource load priority',
      request_info: new Request(`${resource_base}?7`, {importance: 'high'}),
      expected_priority: kHigh
    },
    {
      description: 'fetch() with Request whose importance is "auto" and no request_init must have no effect on resource load priority',
      request_info: new Request(`${resource_base}?8`, {importance: 'auto'}),
      expected_priority: kHigh
    },
    {
      description: 'fetch() with Request whose importance is missing and no request_init must have no effect on resource load priority',
      request_info: new Request(`${resource_base}?9`),
      expected_priority: kHigh
    },
    {
      description: 'fetch() with Request whose importance is "low" and no request_init must be fetched with kLow resource load priority',
      request_info: new Request(`${resource_base}?10`, {importance: 'low'}),
      expected_priority: kLow
    }
  ];

  for (const test of tests) {
    promise_test(async () => {
      const response = await fetch(test.request_info, test.request_init);
      assert_equals(getPriority(response.url), test.expected_priority);
    }, test.description);
  }
</script>
